Ais only select pieces without a cooldown
Why:
Currently the computer players will select pieces to move that have an
active cooldown. This is somewhat distracting to a player and leads to a
less challenging ai.

This commit:
Adds a filtering step in the ai validMoves function to remove all from
pieces that have an active cooldown. Note that this will be run even in
games that don't have the cooldown mechanic. It's possible to separate
this logic out and compose which rules the ai needs to be aware of.
While cleaner, this would certainly be a neglible performace
impromvement so we are not implementing it in this commit for sake of
time.
