+++ src/utils/chess/types.ts
@@ -15,7 +15,7 @@ export interface Empty {
   side: null
   fenId: null
   color: null
-  cooldown?: null
+  cooldown: 0
 }
 
 export interface Piece {
@@ -39,7 +39,7 @@ export interface Piece {
   color: Side
   isPiece: boolean
   fenId: FenId
-  cooldown?: number
+  cooldown: number
 }
 
 /**
